+++ rasa/core/agent.py
@@ -24,7 +24,8 @@ from rasa.core.policies.memoization import MemoizationPolicy
 from rasa.core.processor import MessageProcessor
 from rasa.core.tracker_store import InMemoryTrackerStore
 from rasa.core.trackers import DialogueStateTracker
-from rasa.core.utils import EndpointConfig, LockCounter
+from rasa.utils.endpoints import EndpointConfig
+from rasa.core.utils import LockCounter
 from rasa.nlu.utils import is_url
 
 logger = logging.getLogger(__name__)
@@ -37,8 +38,8 @@ if typing.TYPE_CHECKING:
 
 
 async def load_from_server(
-    agent,
-    model_server: Optional[EndpointConfig] = None
+        agent,
+        model_server: Optional[EndpointConfig] = None
 ) -> 'Agent':
     """Load a persisted model from a server."""
 
@@ -306,10 +307,10 @@ class Agent(object):
                 self.policy_ensemble is not None)
 
     async def handle_message(
-        self,
-        message: UserMessage,
-        message_preprocessor: Optional[Callable[[Text], Text]] = None,
-        **kwargs
+            self,
+            message: UserMessage,
+            message_preprocessor: Optional[Callable[[Text], Text]] = None,
+            **kwargs
     ) -> Optional[List[Text]]:
         """Handle a single message."""
 
@@ -370,10 +371,10 @@ class Agent(object):
 
     # noinspection PyUnusedLocal
     async def log_message(
-        self,
-        message: UserMessage,
-        message_preprocessor: Optional[Callable[[Text], Text]] = None,
-        **kwargs: Any
+            self,
+            message: UserMessage,
+            message_preprocessor: Optional[Callable[[Text], Text]] = None,
+            **kwargs: Any
     ) -> DialogueStateTracker:
         """Append a message to a dialogue - does not predict actions."""
 
@@ -381,12 +382,12 @@ class Agent(object):
         return await processor.log_message(message)
 
     async def execute_action(
-        self,
-        sender_id: Text,
-        action: Text,
-        output_channel: OutputChannel,
-        policy: Text,
-        confidence: float
+            self,
+            sender_id: Text,
+            action: Text,
+            output_channel: OutputChannel,
+            policy: Text,
+            confidence: float
     ) -> DialogueStateTracker:
         """Handle a single message."""
 
@@ -399,11 +400,11 @@ class Agent(object):
                                               confidence)
 
     async def handle_text(
-        self,
-        text_message: Union[Text, Dict[Text, Any]],
-        message_preprocessor: Optional[Callable[[Text], Text]] = None,
-        output_channel: Optional[OutputChannel] = None,
-        sender_id: Optional[Text] = UserMessage.DEFAULT_SENDER_ID
+            self,
+            text_message: Union[Text, Dict[Text, Any]],
+            message_preprocessor: Optional[Callable[[Text], Text]] = None,
+            output_channel: Optional[OutputChannel] = None,
+            sender_id: Optional[Text] = UserMessage.DEFAULT_SENDER_ID
     ) -> Optional[List[Dict[Text, Any]]]:
         """Handle a single message.
 
@@ -727,7 +728,7 @@ class Agent(object):
 
     @staticmethod
     def _create_ensemble(
-        policies: Union[List[Policy], PolicyEnsemble, None]
+            policies: Union[List[Policy], PolicyEnsemble, None]
     ) -> Optional[PolicyEnsemble]:
         if policies is None:
             return None
@@ -746,8 +747,8 @@ class Agent(object):
         """Check whether form policy is present and used."""
 
         has_form_policy = (
-            self.policy_ensemble and
-            any(isinstance(p, FormPolicy)
-                for p in self.policy_ensemble.policies))
+                self.policy_ensemble and
+                any(isinstance(p, FormPolicy)
+                    for p in self.policy_ensemble.policies))
 
         return not self.domain or not self.domain.form_names or has_form_policy
