+++ rasa/utils/__init__.py
@@ -1,0 +1,12 @@
+def configure_colored_logging(loglevel):
+    import coloredlogs
+    field_styles = coloredlogs.DEFAULT_FIELD_STYLES.copy()
+    field_styles['asctime'] = {}
+    level_styles = coloredlogs.DEFAULT_LEVEL_STYLES.copy()
+    level_styles['debug'] = {}
+    coloredlogs.install(
+        level=loglevel,
+        use_chroot=False,
+        fmt='%(asctime)s %(levelname)-8s %(name)s  - %(message)s',
+        level_styles=level_styles,
+        field_styles=field_styles)
