+++ tests/core/test_examples.py
@@ -1,4 +1,3 @@
-import aiohttp
 import sys
 
 import json
@@ -10,7 +9,7 @@ from rasa.core import utils
 from rasa.core.agent import Agent
 from rasa.core.train import train
 from rasa.core.utils import AvailableEndpoints
-from rasa.utils import EndpointConfig, ClientResponseError
+from rasa.utils.endpoints import EndpointConfig, ClientResponseError
 
 
 @pytest.fixture(scope="session")
